server: extract graphql crate
